+++ packages/xchain-crypto/rollup.config.js
@@ -1,6 +1,6 @@
 import typescript from 'rollup-plugin-typescript2'
 import commonjs from 'rollup-plugin-commonjs'
-import json from '@rollup/plugin-json'
+import json from 'rollup-plugin-json'
 import external from 'rollup-plugin-peer-deps-external'
 import resolve from 'rollup-plugin-node-resolve'