+++ packages/xchain-crypto/src/index.ts
@@ -1,4 +1,2 @@
-export * from './secp256k1'
-export * from './ed25519'
 export * from './crypto'
 export * from './utils'